Pre-1.8(V)
Fixed bug where BGI.assignMorphs() was randomly assigning incorrect morphs because it was passing the pruned morph list, rather than the original morph list, to bMorphIsValid(). Since the allowedAttributes section of bMorphIsValid checks the morph list to see if the attribute exists in any of the available morphs, assignMorphs would sometimes prune out morphs that did have that allowedAttribute, leading to the erroneous result that no morphs in the list had that attribute. Debugging code is left here for future use just in case - will be removed in final version.